@@ -14,9 +14,10 @@
     "@types/react-transition-group": "^4.4.3",
     "@types/styled-components": "^5.1.14",
     "axios": "^0.21.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
+    "history": "^5.0.1",
+    "react": "^18.0.0-alpha-8ee4ff883-20211013",
+    "react-dom": "^18.0.0-alpha-8ee4ff883-20211013",
+    "react-router-dom": "^6.0.0-beta.6",
     "react-scripts": "4.0.3",
     "react-transition-group": "^4.4.2",
     "styled-components": "^5.3.1",
